# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/competition": {
            "get": {
                "description": "list competitions",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "List",
                "parameters": [
                    {
                        "type": "string",
                        "description": "competition status, such as done, preparing, in-progress",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "just list competitions of competitor, if it is set",
                        "name": "mine",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CompetitionSummaryDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/competition/{id}": {
            "get": {
                "description": "get detail of competition",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "Get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "competition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.UserCompetitionDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/competition/{id}/competitor": {
            "post": {
                "description": "apply the competition",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "Apply",
                "parameters": [
                    {
                        "description": "body of applying",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_opensourceways_xihe-server_competition_controller.competitorApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/competition/{id}/ranking": {
            "get": {
                "description": "get ranking list of competition",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "GetRankingList",
                "parameters": [
                    {
                        "type": "string",
                        "description": "competition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CompetitonRankingDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/competition/{id}/realted_project": {
            "put": {
                "description": "add related project",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "AddRelatedProject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "competition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "project info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.competitionAddRelatedProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/competition/{id}/submissions": {
            "get": {
                "description": "get submissions",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "GetSubmissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "competition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CompetitionSubmissionsDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            },
            "post": {
                "description": "submit",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "Submit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "competition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "result file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/app.CompetitionSubmissionDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/competition/{id}/team": {
            "get": {
                "description": "get team of competition",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Competition"
                ],
                "summary": "GetTeam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "competition id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.CompetitionTeamDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.CompetitionSubmissionDTO": {
            "type": "object",
            "properties": {
                "project": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "submit_at": {
                    "type": "string"
                }
            }
        },
        "app.CompetitionSubmissionsDTO": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.CompetitionSubmissionDTO"
                    }
                },
                "project": {
                    "type": "string"
                }
            }
        },
        "app.CompetitionSummaryDTO": {
            "type": "object",
            "properties": {
                "bonus": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "desc": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "poster": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "app.CompetitionTeamDTO": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.CompetitionTeamMemberDTO"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "app.CompetitionTeamMemberDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "app.CompetitonRankingDTO": {
            "type": "object",
            "properties": {
                "final": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.RankingDTO"
                    }
                },
                "preliminary": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.RankingDTO"
                    }
                }
            }
        },
...
        "app.RankingDTO": {
            "type": "object",
            "properties": {
                "score": {
                    "type": "number"
                },
                "submit_at": {
                    "type": "string"
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
...
        "app.UserCompetitionDTO": {
            "type": "object",
            "properties": {
                "bonus": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "dataset_doc": {
                    "type": "string"
                },
                "dataset_url": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "doc": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "forum": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_competitor": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "poster": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "team_id": {
                    "type": "string"
                },
                "team_role": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "winners": {
                    "type": "string"
                }
            }
        },
...
        "controller.competitionAddRelatedProjectRequest": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string"
                },
                "project_name": {
                    "type": "string"
                }
            }
        },
...
        "github.com_opensourceways_xihe-server_competition_controller.competitorApplyRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "detail": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string"
                },
                "identity": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                }
            }
        },
...
    }
}